requirements.txt
django==4.2
djangorestframework
Pillow
drf_yasg